+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fixed bug where available serial ports table was not disabled when the port was open.
 
+### Changed
+
+- Removed default menu bar from app.
+
 ## [5.0.4] - 2025-08-02
 
 ### Fixed