+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added global app setting for enabling/disabling automatic updates.
 - Added support for graphing arrays of received data on a single line.
+- Added "Num. bytes in buffer" label to graph view.
 
 ### Fixed